[CID-1032] - Provide service to List user groups for a domain

+++ src/nonCommitTest/groovy/com/rackspace/idm/modules/usergroups/api/resource/DefaultUserGroupCloudServiceRestIntegrationTest.groovy
@@ -1,8 +1,10 @@
 package com.rackspace.idm.modules.usergroups.api.resource
 
 import com.rackspace.docs.identity.api.ext.rax_auth.v1.UserGroup
+import com.rackspace.docs.identity.api.ext.rax_auth.v1.UserGroups
 import com.rackspace.idm.Constants
 import com.rackspace.idm.domain.config.IdentityConfig
+import com.rackspace.idm.domain.service.impl.DefaultDomainService
 import com.rackspace.idm.modules.usergroups.service.DefaultUserGroupService
 import org.apache.commons.lang.RandomStringUtils
 import org.apache.http.HttpStatus
@@ -97,11 +99,12 @@ class DefaultUserGroupCloudServiceRestIntegrationTest extends RootIntegrationTes
         mediaType << [MediaType.APPLICATION_XML_TYPE, MediaType.APPLICATION_JSON_TYPE]
     }
 
-    def "Error check: delete user group"() {
+    @Unroll
+    def "Error check: delete user group; #mediaType"() {
         given:
         UserGroup group = new UserGroup().with {
             it.domainId = sharedUserAdmin.domainId
-            it.name = "addTest_" + RandomStringUtils.randomAlphanumeric(10)
+            it.name = "deleteTest_" + RandomStringUtils.randomAlphanumeric(10)
             it
         }
 
@@ -152,6 +155,107 @@ class DefaultUserGroupCloudServiceRestIntegrationTest extends RootIntegrationTes
 
         then:
         IdmAssert.assertOpenStackV2FaultResponse(deleteResponse, ForbiddenFault, HttpStatus.SC_FORBIDDEN, "Not Authorized")
+
+        cleanup:
+        utils.deleteUsers(users)
+        utils.deleteDomain(domainId)
+        utils.deleteUserGroup(created)
+
+        where:
+        mediaType << [MediaType.APPLICATION_XML_TYPE, MediaType.APPLICATION_JSON_TYPE]
+    }
+
+    @Unroll
+    def "List groups for domain; mediaType = #mediaType"() {
+        when: "Create user group"
+        UserGroup group = new UserGroup().with {
+            it.domainId = sharedUserAdmin.domainId
+            it.name = "listTest_" + RandomStringUtils.randomAlphanumeric(10)
+            it
+        }
+        def response = cloud20.createUserGroup(sharedIdentityAdminToken, group, mediaType)
+
+        then:
+        response.status == HttpStatus.SC_CREATED
+        UserGroup created = response.getEntity(UserGroup)
+
+        when: "List user groups for domain"
+        def getGroupsResponse = cloud20.listUserGroupsForDomain(sharedIdentityAdminToken, sharedUserAdmin.domainId, null, mediaType)
+        UserGroups groups = getGroupsResponse.getEntity(UserGroups)
+
+        then:
+        getGroupsResponse.status == HttpStatus.SC_OK
+        groups.userGroup.size() > 0
+        groups.userGroup.find {it.name == created.name} != null
+
+        when: "List user groups for domain with name query param"
+        getGroupsResponse = cloud20.listUserGroupsForDomain(sharedIdentityAdminToken, sharedUserAdmin.domainId, group.name, mediaType)
+        groups = getGroupsResponse.getEntity(UserGroups)
+
+        then:
+        getGroupsResponse.status == HttpStatus.SC_OK
+
+        and:
+        groups.userGroup.size() == 1
+        UserGroup userGroup = groups.userGroup.get(0)
+        userGroup.id == created.id
+        userGroup.name == created.name
+        userGroup.domainId == created.domainId
+        userGroup.description == created.description
+
+        cleanup:
+        utils.deleteUserGroup(created)
+
+        where:
+        mediaType << [MediaType.APPLICATION_XML_TYPE, MediaType.APPLICATION_JSON_TYPE]
+    }
+
+    @Unroll
+    def "Error check: list user groups for domain; #mediaType"() {
+        given:
+        UserGroup group = new UserGroup().with {
+            it.domainId = sharedUserAdmin.domainId
+            it.name = "listTest_" + RandomStringUtils.randomAlphanumeric(10)
+            it
+        }
+
+        def domainId = utils.createDomain()
+        def defaultUser, users
+        (defaultUser, users) = utils.createDefaultUser(domainId)
+
+        when: "Create user group"
+        def response = cloud20.createUserGroup(sharedIdentityAdminToken, group, mediaType)
+        UserGroup created = response.getEntity(UserGroup)
+
+        then:
+        response.status == HttpStatus.SC_CREATED
+
+        when: "Invalid domain"
+        def getGroupsResponse = cloud20.listUserGroupsForDomain(sharedIdentityAdminToken, "invalid")
+        String errMsg = String.format(DefaultDomainService.DOMAIN_NOT_FOUND_ERROR_MESSGAE, "invalid")
+
+        then:
+        IdmAssert.assertOpenStackV2FaultResponse(getGroupsResponse, ItemNotFoundFault, HttpStatus.SC_NOT_FOUND, errMsg)
+
+        when: "Invalid name query param"
+        getGroupsResponse = cloud20.listUserGroupsForDomain(sharedIdentityAdminToken, sharedUserAdmin.domainId, "invalid")
+        UserGroups userGroups = getGroupsResponse.getEntity(UserGroups)
+
+        then:
+        userGroups.userGroup.size() == 0
+
+        when: "Invalid auth token"
+        getGroupsResponse = cloud20.listUserGroupsForDomain("invalid", created.domainId)
+        errMsg = "No valid token provided. Please use the 'X-Auth-Token' header with a valid token."
+
+        then:
+        IdmAssert.assertOpenStackV2FaultResponse(getGroupsResponse, UnauthorizedFault, HttpStatus.SC_UNAUTHORIZED, errMsg)
+
+        when: "Unauthorized token"
+        getGroupsResponse = cloud20.listUserGroupsForDomain(utils.getToken(defaultUser.username), created.domainId)
+
+        then:
+        IdmAssert.assertOpenStackV2FaultResponse(getGroupsResponse, ForbiddenFault, HttpStatus.SC_FORBIDDEN, "Not Authorized")
 
         cleanup:
         utils.deleteUsers(users)
